Add Level-1, Echo
Chatbot only displays welcome message.

Echo method is required within chatbot.

Echo method created which echoes text from users. It exits when string bye, case insensitive is detected.
